Refactor data to handle depth 3+ texts

+++ scripts/JsonExporterForIOS.py
@@ -68,6 +68,8 @@ def zip_last_text(title):
 	z = zipfile.ZipFile(zipPath, "w", zipfile.ZIP_DEFLATED)
 
 	for file in glob.glob("*.json"):
+		if file.endswith("calendar.json") or file.endswith("toc.json"):
+			continue
 		z.write(file)
 		os.remove(file)
 	z.close()
@@ -82,15 +84,19 @@ def export_texts(skip_existing=False):
 	titles = [i.title for i in model.library.all_index_records(with_commentary=True)]
 
 	for title in titles:
-		if skip_existing:
-			if os.path.isfile("%s/%s.zip" % (SEFARIA_EXPORT_PATH, title)):
-				continue 
+		if skip_existing and os.path.isfile("%s/%s.zip" % (SEFARIA_EXPORT_PATH, title)):
+			continue 
 		export_text(title)
-		export_index(title)
-		zip_last_text(title)
 
 
 def export_text(title):
+	"""Writes a ZIP file containing text content json and text index JSON"""
+	export_text_json(title)
+	export_index(title)
+	zip_last_text(title)	
+
+
+def export_text_json(title):
 	"""
 	Takes a single document from the `texts` collection exports it, by chopping it up 
 	Add helpful data like 
@@ -98,67 +104,64 @@ def export_text(title):
 	print title
 	try:
 		for oref in model.get_index(title).all_top_section_refs():
-			text = model.TextFamily(oref, version=None, lang=None, commentary=0, context=0, pad=0, alts=False).contents()
-			text["next"]	= oref.next_section_ref().normal() if oref.next_section_ref() else None
-			text["prev"]	= oref.prev_section_ref().normal() if oref.prev_section_ref() else None
-			text["content"] = []
-			
-			if str(oref) == "Sha'ar Ha'Gemul of the Ramban 1":
-				print "Sha'ar Ha'Gemul of the Ramban 1 is the worst"
+			if oref.is_section_level():
+				doc = section_data(oref)
 			else:
-				for x in range (0,max([len(text["text"]),len(text["he"])])):
-					curContent = {}
-					curContent["segmentNumber"] = str(x+1)
-
-					links = get_links(text["ref"]+":"+curContent["segmentNumber"], False)
-					for link in links:
-						del link['commentator']
-						del link['heCommentator']
-						del link['type']
-						del link['anchorText']
-						del link['commentaryNum']
-						if 'heTitle' in link: del link['heTitle']
-						del link['_id']
-						del link['anchorRef']
-						del link['ref']
-						del link['anchorVerse']
+				sections = oref.all_subrefs()
+				doc = {
+					"ref": oref.normal(),
+					"sections": {}
+				}
+				for section in sections:
+					doc["sections"][section.normal()] = section_data(section)
 					
-					curContent["links"] = links
-				
-					if x < len(text["text"]): curContent["text"]=text["text"][x]
-					else: curContent["text"]=""
-					if x < len(text["he"]): curContent["he"]=text["he"][x]
-					else: curContent["he"]=""
+			path = make_path(doc, "json")
+			write_doc(doc, path)
 
-					text["content"].append(curContent)
-
-				text.pop("maps", None)
-				text.pop("versionSource", None)
-				text.pop("heDigitizedBySefaria", None)
-				text.pop("heVersionTitle", None)
-				text.pop("heVersionNotes", None)
-				text.pop("heVersionStatus", None)
-				text.pop("isSpanning", None)
-				text.pop("heVersionSource", None)
-				text.pop("versionNotes", None)
-				text.pop("versionTitle", None)
-				text.pop("heLicense", None)
-				text.pop("digitizedBySefaria", None)
-				text.pop("versions", None)
-				text.pop("license", None)
-				text.pop("versionStatus", None)
-				text.pop("heSources", None)
-				text.pop("sources", None)
-				text.pop("he",None)
-				text.pop("text",None)
-				
-				path = make_path(text, "json")
-				write_doc(text, path)
 	except Exception, e:
 		print "Error exporting %s: %s" % (title, e)
 		import traceback
 		print traceback.format_exc()
-			
+
+
+def section_data(oref):
+	"""
+	Returns a dictionary with all the data we care about for section level `oref`.
+	"""
+	text = model.TextFamily(oref, version=None, lang=None, commentary=0, context=0, pad=0, alts=False).contents()
+	data = {
+		"ref": text["ref"],
+		"heRef": text["heRef"],
+		"indexTitle": text["indexTitle"],
+		"heTitle": text["heTitle"],
+		"sectionRef": text["sectionRef"],
+		"next":	oref.next_section_ref().normal() if oref.next_section_ref() else None,
+		"prev": oref.prev_section_ref().normal() if oref.prev_section_ref() else None,
+		"content": [],
+	}
+	
+	for x in range (0,max([len(text["text"]),len(text["he"])])):
+		curContent = {}
+		curContent["segmentNumber"] = str(x+1)
+
+		links = get_links(text["ref"] + ":" + curContent["segmentNumber"], False)
+		def simple_link(link):
+			return {
+				"category":    link["category"],
+				"sourceHeRef": link["sourceHeRef"],
+				"index_title": link["index_title"],
+				"sourceRef":   link["sourceRef"]
+			}
+		if len(links) > 0:
+			curContent["links"] = [simple_link(link) for link in links]
+	
+		if x < len(text["text"]): curContent["text"]=text["text"][x]
+		if x < len(text["he"]): curContent["he"]=text["he"][x]
+
+		data["content"].append(curContent)
+
+	return data
+
 
 def export_index(title):
 	"""
